完善 mock 数据方式
